feat: add support for format in data type for object properties
- added support for mapping format to language specific data type in generic generator
- added typing for property data type & its decoder
- migrated configuration & execution constants to runtime static module
- added prettierrc, prettierignore & editorconfig for formatting
- formatted all files with new config

+++ src/writer/index.ts
@@ -1,13 +1,14 @@
-import { Configuration, GenerationResult, TypesOutput } from "$types";
+import { Configuration, GenerationResult, TypesOutput } from '$types';
 import {
   addValuesToMappedSet,
   createFolder,
   createFolderWithBasePath,
   deleteFolder,
   getCompleteFolderPath,
-  writeFile,
-} from "$utils";
-import Handlebars from "handlebars";
+  writeFile
+} from '$utils';
+import Handlebars from 'handlebars';
+import { Runtime } from '../runtime';
 
 //#region Localized types
 
@@ -21,11 +22,11 @@ type FileWriterOutput = {
 async function writeTypesToFiles(
   config: Configuration,
   types: TypesOutput,
-  folderName: string = "",
+  folderName: string = ''
 ): Promise<FileWriterOutput> {
   const result: FileWriterOutput = {
-    folderName: config.output.directory + "/" + folderName,
-    files: [],
+    folderName: config.output.directory + '/' + folderName,
+    files: []
   };
   for (let typeName in types) {
     const file = typeName + config.output.fileExtension;
@@ -38,27 +39,25 @@ async function writeTypesToFiles(
 async function writeTypesToFile(
   config: Configuration,
   types: TypesOutput,
-  fileName: string = "types",
+  fileName: string = 'types'
 ): Promise<FileWriterOutput> {
-  let content = "";
+  let content = '';
   for (let typeName in types) {
-    content += types[typeName] + "\n";
+    content += types[typeName] + '\n';
   }
-  await writeFile(
-    config.output.directory,
-    fileName + config.output.fileExtension,
-    content,
-  );
+  await writeFile(config.output.directory, fileName + config.output.fileExtension, content);
   return {
     folderName: config.output.directory,
-    files: [fileName + config.output.fileExtension],
+    files: [fileName + config.output.fileExtension]
   };
 }
 
-export async function writeOutput(
-  config: Configuration,
-  generationResult: GenerationResult,
-) {
+export async function writeOutput(generationResult: GenerationResult) {
+  if (Runtime.config === null) {
+    throw new Error('Configuration not set!');
+  }
+  const config = Runtime.config;
+
   if (config.output.cleanWrite) {
     await deleteFolder(config.output.directory);
   }
@@ -70,75 +69,69 @@ export async function writeOutput(
 
   let typesFilesWritten = null;
 
-  if (config.output.writerMode.types === "Files") {
+  if (config.output.writerMode.types === 'Files') {
     typesFilesWritten = await writeTypesToFiles(config, generationResult.types);
-  } else if (config.output.writerMode.types === "SingleFile") {
+  } else if (config.output.writerMode.types === 'SingleFile') {
     typesFilesWritten = await writeTypesToFile(config, generationResult.types);
   }
   if (typesFilesWritten !== null) {
     addValuesToMappedSet(
       writtenFiles,
       await getCompleteFolderPath(typesFilesWritten.folderName),
-      typesFilesWritten.files,
+      typesFilesWritten.files
     );
   }
 
   // writing grouped types to output directory
-  if (config.output.writerMode.groupedTypes === "FolderWithFiles") {
+  if (config.output.writerMode.groupedTypes === 'FolderWithFiles') {
     for (let groupName in generationResult.groupedTypes) {
       let groupFilesWritten = null;
       await createFolderWithBasePath(config.output.directory, groupName);
-      addValuesToMappedSet(
-        writtenFiles,
-        await getCompleteFolderPath(config.output.directory),
-        [groupName],
-      );
+      addValuesToMappedSet(writtenFiles, await getCompleteFolderPath(config.output.directory), [
+        groupName
+      ]);
       groupFilesWritten = await writeTypesToFiles(
         config,
         generationResult.groupedTypes[groupName],
-        groupName,
+        groupName
       );
       if (groupFilesWritten !== null) {
         addValuesToMappedSet(
           writtenFiles,
           await getCompleteFolderPath(groupFilesWritten.folderName),
-          groupFilesWritten.files,
+          groupFilesWritten.files
         );
       }
     }
-  } else if (config.output.writerMode.groupedTypes === "SingleFile") {
+  } else if (config.output.writerMode.groupedTypes === 'SingleFile') {
     for (let groupName in generationResult.groupedTypes) {
       let groupFilesWritten = null;
       groupFilesWritten = await writeTypesToFile(
         config,
         generationResult.groupedTypes[groupName],
-        groupName,
+        groupName
       );
       if (groupFilesWritten !== null) {
         addValuesToMappedSet(
           writtenFiles,
           await getCompleteFolderPath(groupFilesWritten.folderName),
-          groupFilesWritten.files,
+          groupFilesWritten.files
         );
       }
     }
   }
 
   // writing exporter modules
-  const exporterModuleTemplate = Handlebars.compile(
-    config.template.exporterModuleSyntax,
-  );
+  const exporterModuleTemplate = Handlebars.compile(config.template.exporterModuleSyntax);
 
   writtenFiles.forEach(async (files, folder) => {
     const exporterModuleContent = exporterModuleTemplate({
-      modules: [...files].map((file) =>
-        file.replace(config.output.fileExtension, ""),
-      ),
+      modules: [...files].map((file) => file.replace(config.output.fileExtension, ''))
     });
     await writeFile(
       folder,
       config.language.exporterModuleName + config.output.fileExtension,
-      exporterModuleContent,
+      exporterModuleContent
     );
   });
 }
